+++ pkg/client/registry.go
@@ -21,6 +21,8 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
+const DefaultPublicRegistryURL = "https://registry.terraform.io"
+
 // createHTTPClient initializes a retryable HTTP client
 func createHTTPClient(insecureSkipVerify bool, logger *log.Logger) *http.Client {
 	retryClient := retryablehttp.NewClient()
@@ -66,7 +68,7 @@ func SendRegistryCall(client *http.Client, method string, uri string, logger *lo
 		ver = callOptions[0] // API version will be the first optional arg to this function
 	}
 
-	url, err := url.Parse(fmt.Sprintf("https://registry.terraform.io/%s/%s", ver, uri))
+	url, err := url.Parse(fmt.Sprintf("%s/%s/%s", DefaultPublicRegistryURL, ver, uri))
 	if err != nil {
 		return nil, fmt.Errorf("error parsing terraform registry URL: %w", err)
 	}